pyproject.toml
[tool.poetry]
name = "voicebot"
version = "0.0.0"
description = "A simple Danish voice bot."
authors = [
    "Dan Saattrup Nielsen <dan.nielsen@alexandra.dk>",
]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
hydra-core = "^1.3.2"
pyaudio = "^0.2.13"
numpy = "^1.26.0"
transformers = "^4.34.0"
torch = "^2.1.0"
scipy = "^1.11.3"
openai = "^0.28.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-xdist = "^3.2.1"
pytest-cov = "^4.0.0"
pre-commit = "^3.3.1"
pdoc = "^13.1.1"
black = "^23.3.0"
isort = "^5.12.0"
readme-coverage-badger = ">=0.1.2,<1.0.0"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    '--verbose',
    '--durations=10',
    '--color=yes',
    '-s',
    '-vv',
    '--doctest-modules',
    '--cov=src/voicebot',
    '-n 2',
]
xfail_strict = true
filterwarnings = ["ignore"]
log_cli_level = "info"
testpaths = ["tests", "src/voicebot"]

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''

[tool.ruff]
target-version = "py311"

[tool.ruff.extend-per-file-ignores]
"__init__.py" = ["F401"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
